Add SASPlanet page UI protype

+++ EasyGrid/ViewModels/CoordinatesCreateViewModel.cs
@@ -6,29 +6,29 @@ namespace EasyGrid.ViewModels
 {
     internal class CoordinatesCreateViewModel : ObservableObject
     {
-        private int _point1Lat;
-        public int Point1Lat
+        private double _point1Lat;
+        public double Point1Lat
         {
             get => _point1Lat;
             set => SetProperty(ref _point1Lat, value);
         }
 
-        private int _point1Lon;
-        public int Point1Lon
+        private double _point1Lon;
+        public double Point1Lon
         {
             get => _point1Lon;
             set => SetProperty(ref _point1Lon, value);
         }
 
-        private int _point9Lat;
-        public int Point9Lat
+        private double _point9Lat;
+        public double Point9Lat
         {
             get => _point9Lat;
             set => SetProperty(ref _point9Lat, value);
         }
 
-        private int _point9Lon;
-        public int Point9Lon
+        private double _point9Lon;
+        public double Point9Lon
         {
             get => _point9Lon;
             set => SetProperty(ref _point9Lon, value);
